@@ -31,7 +31,7 @@ public class HighArrayApp {
 		if (arr.find(searchKey))
 			System.out.println("Found " + searchKey);
 		else
-			System.out.println("Can�t find " + searchKey);
+			System.out.println("Can't find " + searchKey);
 		arr.delete(00); // delete 3 items
 		arr.delete(55);
 		arr.delete(99);
