@@ -1331,7 +1331,7 @@ public class TreeList<E extends Comparable<? super E>> {
 	 * 1. Initialize current as root 
 	 * 2. While current is not NULL
 	 * If current does not have left child
-	 * a) Print current�s data
+	 * a) Print current's data
 	 * b) Go to the right, i.e., current = current->right
 	 * Else
 	 * a) Make current as right child of the rightmost node in current's left subtree
@@ -1443,7 +1443,7 @@ public class TreeList<E extends Comparable<? super E>> {
 	 * 1. Initialize current as root 
 	 * 2. While current is not NULL
 	 * If current does not have left child
-	 * a) Print current�s data
+	 * a) Print current's data
 	 * b) Go to the right, i.e., current = current->right
 	 * Else
 	 * a) Make current as right child of the rightmost node in current's left subtree
